ISSUE#76 MISSING UPDATES:
Added an out of date check to jenkins fetcher so that the build numbers
are checked and only if the last is one less than the building will
itnot update. This should catch updates on launch and failures that are
restarted immediately. Added field to jenkins job for current build
number. Added import for build number in build state check. Fixed issue
where job description panel was looking at the previous build number not
the current.