ISSUE#26 PERSISTENCE:
Implemented sessions configuration for the build wall and integrated
with dual build wall. Using the JUPA library, each configuration (left
and right) are persisted to files at the same location as the jar. In
this build only an automatic feature is available for persistence. Also
provided synchronized observable map to get around a weakness in
JavaFx's equivalent.