ISSUE#103 AUTO UPDATE MECHANISM:
Software changes while waiting for next JUPA build to be released. The
build wall has been wrapped with a release panel that provides a
notification when a new version is available. It has been configured to
point to a file in the repo with the releases that can be updated to.