THEME CONFIGURATION HOOK UP:
Added theme configuration to configuration tree so that it is usable but
not yet linked to the actualy build wall or saving. Extracted scroll
pane from configuration pane so that individual content panels can
define how it is to be used, with a base implementation to share the
behaviour. Added some syling elements such as hiding a scroll panes
background.