ISSUE#119 WIFI DISCONNECTING:
When testing locally an issue occurs when the api cannot get a response,
therefore null is the response, which is then parsed. This naturally
fails. The Api message converter is defined to return nulls but the
polling doesn't handle these nulls. This has been fixed with a simple
null check which is valid because the detected nulls are correctly
reported, so the thread is able to continue polling. Not sure if this is
the only issue however. Resolving this issue because this is all that is
known, further issues can get a more updated issue.