BUILD RESULT EARLY UPDATE ISSUE:
An issue was present where jenkins would appear to notify the status
change before the build was complete - or certain properties were
populated - this caused builds that had changed to show no change in
some circumstances (since the status had updated just prior to the
identification of the completion, so at the time of completion there was
no change). This has hopefully been resolved by immediately notifying
whenever the status changes and then only notifying if not already.